+++ test/integration/pruning_test.exs
@@ -13,7 +13,7 @@ defmodule Oban.Integration.PruningTest do
     %Job{id: _id_} = insert_job(state: "completed")
     %Job{id: id_6} = insert_job(state: "completed")
 
-    start_supervised!({Oban, repo: Repo, prune: {:maxlen, 1}})
+    start_supervised!({Oban, repo: Repo, prune: {:maxlen, 1}, prune_interval: 10})
 
     with_backoff(fn ->
       assert retained_ids() == [id_1, id_4, id_6]
@@ -28,7 +28,7 @@ defmodule Oban.Integration.PruningTest do
     %Job{id: id_1} = insert_job(state: "completed", completed_at: minutes_ago(3))
     %Job{id: id_2} = insert_job(state: "completed", completed_at: minutes_ago(1))
 
-    start_supervised!({Oban, repo: Repo, prune: {:maxage, 60 * 4}})
+    start_supervised!({Oban, repo: Repo, prune: {:maxage, 60 * 4}, prune_interval: 10})
 
     with_backoff(fn ->
       assert retained_ids() == [id_1, id_2]